third_party/WebKit/LayoutTests/external/wpt/xhr/abort-during-upload-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: abort() while sending data assert_equals: expected "upload.loadstart(0,9999,true)" but got "upload.loadstart(0,0,false)"
Harness: the test ran to completion.
